CheckReservationのコンポーネント化
